+++ data-transfer-object/src/main/java/org/grnet/status/dtos/topology/WebApiFeedsTopologyResponse.java
@@ -1,0 +1,15 @@
+package org.grnet.status.dtos.topology;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.grnet.status.dtos.Status;
+
+import java.util.List;
+
+public class WebApiFeedsTopologyResponse {
+
+    @JsonProperty("status")
+    public Status status;
+
+    @JsonProperty("data")
+    public List<FeedTopologyDto> data;
+}